Module fuz_json_parser::json_parser

Modules

states

Functions

main_parse
the recursive main section of the parsing algorithm given any properly formated json string or subset value it will return the corisponding JsonValue type

parse
the main parsing method entry point returns the parsed data as a JsonValue or it errors and returns an error message

parse_block
parses a block of the json string the block is either an array or an object

parse_key_pair
parse a key value pair from AsRef<str> parse out the String key and JsonValue value

parse_single_value
parse a single value from an AsRef<str> to JsonValue

Type Definitions

Result